Revert "Applying SRP (1) SOLID Principle and clean code"
This reverts commit dbb6f93ae2b5e1c593c9d154bc233cdd5c3ff0ec.
